Additional improvements
Improved GuiContributors:
* automatically selects first entry when opened
* navigate with arrows, tab, shift+tab, esc
* simplified with kotlin getter, ...

Improved AltManager (similar to GuiContributors)
* simplified navigation code
* added tab, shift+tab support

* Simplified floats, removed unnecessary .toFloat(), ...

* Additional stuff similar to previous changes... (imported static methods and fields from GlStateManager, RenderUtils, ...)

+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/player/NoFall.kt
@@ -133,7 +133,7 @@ class NoFall : Module() {
                         mc.thePlayer.posX, Double.NaN, mc.thePlayer.posZ, false
                     )
                 )
-                mc.thePlayer.fallDistance = (-9999).toFloat()
+                mc.thePlayer.fallDistance = -9999f
             }
             "spartan" -> {
                 spartanTimer.update()
